gcc-4.9: correct symlinks
This is needed because the actual x86_64-linux-android-{gcc,g++} is a python wrapper that prepends real- to its own filename. However, the real-x86_64-linux-android-{gcc,g++} is also a wrapper that does the same thing, and so you get real-real-x86_64-linux-android-{gcc,g++}. This patch corrects that by pointing real-x86_64-linux-android-{gcc,g++} to the actual compiler.

Change-Id: Ib7a86c01d04d3563da46178c043fd9599c88f1b5
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>